+++ src/main/java/uk/gov/ons/census/casesvc/service/CcsToFieldService.java
@@ -38,7 +38,7 @@ public class CcsToFieldService {
     ccsToField.setLatitude(caze.getLatitude());
     ccsToField.setLongitude(caze.getLongitude());
     ccsToField.setCaseId(caze.getCaseId().toString());
-    ccsToField.setCaseRef(Integer.toString(caze.getCaseRef()));
+    ccsToField.setCaseRef(Long.toString(caze.getCaseRef()));
     ccsToField.setAddressType(caze.getAddressType());
     ccsToField.setFieldCoordinatorId(caze.getFieldCoordinatorId());
     ccsToField.setUndeliveredAsAddress(false);
